+++ Boa.Constrictor.UnitTests/RestSharp/Serialization/RequestDataTest.cs
@@ -12,7 +12,28 @@ namespace Boa.Constrictor.UnitTests.RestSharp
     public class RequestDataTest
     {
         [Test]
-        public void Init()
+        public void Init_Basic()
+        {
+            var clientUri = new Uri("https://www.pl.com");
+            var resource = "/path/to/thing";
+            var requestMethod = Method.GET;
+
+            var clientMock = new Mock<IRestClient>();
+            clientMock.Setup(x => x.BaseUrl).Returns(clientUri);
+            clientMock.Setup(x => x.BuildUri(It.IsAny<IRestRequest>())).Returns(clientUri);
+
+            var request = new RestRequest(resource, requestMethod);
+
+            var data = new RequestData(request, clientMock.Object);
+            data.Method.Should().Be(requestMethod.ToString());
+            data.Uri.Should().Be(clientUri.ToString());
+            data.Resource.Should().Be(resource);
+            data.Parameters.Should().BeEmpty();
+            data.Body.Should().BeNull();
+        }
+
+        [Test]
+        public void Init_Full()
         {
             var clientUri = new Uri("https://www.pl.com");
             var resource = "/path/to/thing";
@@ -24,17 +45,19 @@ namespace Boa.Constrictor.UnitTests.RestSharp
                 new Parameter("p2", "goodbye", ParameterType.Cookie),
             };
 
+            var body = new RequestBody("json", "body", "value");
+
             var clientMock = new Mock<IRestClient>();
             clientMock.Setup(x => x.BaseUrl).Returns(clientUri);
             clientMock.Setup(x => x.BuildUri(It.IsAny<IRestRequest>())).Returns(clientUri);
 
             var requestMock = new Mock<IRestRequest>();
-            requestMock = new Mock<IRestRequest>();
             requestMock.Setup(x => x.Method).Returns(requestMethod);
             requestMock.Setup(x => x.Resource).Returns(resource);
             requestMock.Setup(x => x.Parameters).Returns(parameters);
+            requestMock.Setup(x => x.Body).Returns(body);
 
-            var data = new RequestData(clientMock.Object, requestMock.Object);
+            var data = new RequestData(requestMock.Object, clientMock.Object);
             data.Method.Should().Be(requestMethod.ToString());
             data.Uri.Should().Be(clientUri.ToString());
             data.Resource.Should().Be(resource);
@@ -45,6 +68,7 @@ namespace Boa.Constrictor.UnitTests.RestSharp
             data.Parameters[1].Name.Should().Be("p2");
             data.Parameters[1].Value.Should().Be("goodbye");
             data.Parameters[1].Type.Should().Be(ParameterType.Cookie.ToString());
+            data.Body.Should().Be(body);
         }
     }
 }