+++ Boa.Constrictor/WebDriver/Extensions/AbstractWebQuestion.cs
@@ -1,5 +1,6 @@
 ﻿using Boa.Constrictor.Screenplay;
 using OpenQA.Selenium;
+using System;
 
 namespace Boa.Constrictor.WebDriver
 {
@@ -30,7 +31,20 @@ namespace Boa.Constrictor.WebDriver
         /// <param name="actor">The screenplay actor.</param>
         /// <returns></returns>
         public virtual TAnswer RequestAs(IActor actor) => RequestAs(actor, actor.Using<BrowseTheWeb>().WebDriver);
-        
+
+        /// <summary>
+        /// Checks if this interaction is equal to another interaction.
+        /// </summary>
+        /// <param name="obj">The other object.</param>
+        /// <returns></returns>
+        public override bool Equals(object obj) => obj.GetType().Equals(GetType());
+
+        /// <summary>
+        /// Gets a unique hash code for this interaction.
+        /// </summary>
+        /// <returns></returns>
+        public override int GetHashCode() => HashCode.Combine(GetType());
+
         /// <summary>
         /// Returns a description of the question.
         /// </summary>