+++ Boa.Constrictor/WebDriver/Tasks/AcceptAlert.cs
@@ -1,5 +1,6 @@
 ﻿using Boa.Constrictor.Screenplay;
 using OpenQA.Selenium;
+using System;
 
 namespace Boa.Constrictor.WebDriver
 {
@@ -65,6 +66,21 @@ namespace Boa.Constrictor.WebDriver
                     throw;
             }
         }
+
+        /// <summary>
+        /// Checks if this interaction is equal to another interaction.
+        /// </summary>
+        /// <param name="obj">The other object.</param>
+        /// <returns></returns>
+        public override bool Equals(object obj) =>
+            obj is AcceptAlert interaction &&
+            RethrowNoAlert == interaction.RethrowNoAlert;
+
+        /// <summary>
+        /// Gets a unique hash code for this interaction.
+        /// </summary>
+        /// <returns></returns>
+        public override int GetHashCode() => HashCode.Combine(GetType(), RethrowNoAlert);
 
         /// <summary>
         /// Returns a description of the task.