+++ Boa.Constrictor/WebDriver/Tasks/ScrollToElement.cs
@@ -1,5 +1,7 @@
 ﻿using Boa.Constrictor.Screenplay;
 using OpenQA.Selenium;
+using System;
+using System.Collections.Generic;
 
 namespace Boa.Constrictor.WebDriver
 {
@@ -62,6 +64,22 @@ namespace Boa.Constrictor.WebDriver
         /// <param name="driver">The WebDriver.</param>
         public override void PerformAs(IActor actor, IWebDriver driver) =>
             actor.Calls(JavaScript.On(Locator, ScrollCommand));
+
+        /// <summary>
+        /// Checks if this interaction is equal to another interaction.
+        /// </summary>
+        /// <param name="obj">The other object.</param>
+        public override bool Equals(object obj) =>
+            obj is ScrollToElement element &&
+            EqualityComparer<IWebLocator>.Default.Equals(Locator, element.Locator) &&
+            AlignToTop == element.AlignToTop;
+
+        /// <summary>
+        /// Gets a unique hash code for this interaction.
+        /// </summary>
+        /// <returns></returns>
+        public override int GetHashCode() => 
+            HashCode.Combine(GetType(), Locator, AlignToTop);
 
         /// <summary>
         /// Returns a description of the task.