+++ Boa.Constrictor/WebDriver/Tasks/Select.cs
@@ -1,6 +1,8 @@
 ﻿using Boa.Constrictor.Screenplay;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Support.UI;
+using System;
+using System.Collections.Generic;
 
 namespace Boa.Constrictor.WebDriver
 {
@@ -114,6 +116,25 @@ namespace Boa.Constrictor.WebDriver
                 throw new BrowserInteractionException(
                     $"No select method (index, text, or value) provided for Select task");
         }
+
+        /// <summary>
+        /// Checks if this interaction is equal to another interaction.
+        /// </summary>
+        /// <param name="obj">The other object.</param>
+        public override bool Equals(object obj) =>
+            obj is Select select &&
+            EqualityComparer<IWebLocator>.Default.Equals(Locator, select.Locator) &&
+            Index == select.Index &&
+            PartialMatch == select.PartialMatch &&
+            Text == select.Text &&
+            Value == select.Value;
+
+        /// <summary>
+        /// Gets a unique hash code for this interaction.
+        /// </summary>
+        /// <returns></returns>
+        public override int GetHashCode() =>
+            HashCode.Combine(GetType(), Locator, $"Index {Index}", $"PartialMatch {PartialMatch}", $"Text {Text}", $"Value {Value}");
 
         #endregion
     }